+++ c/zss.c
@@ -1259,6 +1259,23 @@ static bool readAgentHttpsSettingsV2(ShortLivedHeap *slh,
   if (getStatus){
     zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_SEVERE, "internal error accessing .components.zss.port, err=%d\n", getStatus);
   }
+  char *verifyCertificates = NULL;
+  if ((getStatus = cfgGetStringC(configmgr, ZSS_CFGNAME, &verifyCertificates, 2, "zowe", "verifyCertificates")) != 0) {
+    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_WARNING, "internal error accessing .zowe.verifyCertificates, err=%d, will default to STRICT\n", getStatus);
+  }
+  if (!verifyCertificates || 0 == strcasecmp(verifyCertificates, "STRICT")) {
+    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_INFO, "verifyCertificates mode: STRICT\n");
+    settings->certVerify = TLS_CERTVERIFY_STRICT;
+  } else if (0 == strcasecmp(verifyCertificates, "NONSTRICT")) {
+    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_INFO, "verifyCertificates mode: NONSTRICT\n");
+    settings->certVerify = TLS_CERTVERIFY_NONSTRICT;
+  } else if (0 == strcasecmp(verifyCertificates, "DISABLED")) {
+    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_INFO, "verifyCertificates mode: DISABLED\n");
+    settings->certVerify = TLS_CERTVERIFY_DISABLED;
+  } else {
+    zowelog(NULL, LOG_COMP_ID_MVD_SERVER, ZOWE_LOG_WARNING, ".zowe.verifyCertificates got an unknown value \"%s\", will default to STRICT\n", verifyCertificates);
+    settings->certVerify = TLS_CERTVERIFY_STRICT;
+  }
   bool useTls = false;
   cfgGetBooleanC(configmgr,ZSS_CFGNAME,&useTls,3,"components","zss","tls");
   bool isHttpsConfigured = useTls && settings->keyring;